requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
distro==1.9.0
Django==5.0.4
django-allauth==0.61.1
djangorestframework==3.15.1
exceptiongroup==1.2.1
firebase-admin==6.5.0
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.4.1
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
msgpack==1.0.8
oauthlib==3.2.2
openai==1.23.5
pexels-api==1.0.1
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==2.8.0
pyparsing==3.1.2
python-decouple==3.8
python-multipart==0.0.9
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.2
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
